01-Inventory-Management/05-product-catalogue/requirements_seo.txt
pyairtable
google-generativeai
python-dotenv
python-wordpress-xmlrpc
tldextract